requirements.txt
asgiref==3.5.2
asn1crypto==1.5.1
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
dj-database-url==1.2.0
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flake8==6.0.0
idna==3.4
iniconfig==1.1.1
install==1.3.5
mccabe==0.7.0
packaging==22.0
pg8000==1.29.4
pluggy==1.0.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
pytest==7.2.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
scramp==1.4.4
six==1.16.0
sqlparse==0.4.3
testing.common.database==2.0.3
testing.postgresql==1.3.0
urllib3==1.26.13
